agents/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
httpx==0.28.1
openai==1.57.0
langchain==0.3.10
langchain-openai==0.2.10
python-multipart==0.0.20
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0